+++ lib/adhearsion/call_controller.rb
@@ -238,7 +238,7 @@ module Adhearsion
       block_until_resumed
       async = (target.is_a?(Hash) ? target : options).delete :async
       join_command = call.join target, options
-      waiter = join_command.call_id || join_command.mixer_name
+      waiter = join_command.call_uri || join_command.mixer_name
       if async
         call.wait_for_joined waiter
       else