+++ lib/adhearsion/punchblock_plugin/initializer.rb
@@ -5,7 +5,7 @@ require 'blather'
 module Adhearsion
   class PunchblockPlugin
     class Initializer
-      cattr_accessor :config, :client, :connection, :dispatcher, :attempts
+      cattr_accessor :config, :client, :dispatcher, :attempts
 
       self.attempts = 0
 
@@ -14,16 +14,10 @@ module Adhearsion
           self.config = Adhearsion.config[:punchblock]
 
           username = self.config.username
-          connection_class = case (self.config.platform || :xmpp)
-          when :xmpp
+          if (self.config.platform || :xmpp) == :xmpp
             username = Blather::JID.new username
             username = Blather::JID.new username.node, username.domain, resource unless username.resource
             username = username.to_s
-            Punchblock::Connection::XMPP
-          when :asterisk
-            Punchblock::Connection::Asterisk
-          when :freeswitch
-            Punchblock::Connection::Freeswitch
           end
 
           connection_options = {
@@ -40,8 +34,7 @@ module Adhearsion
             :default_voice      => self.config.default_voice
           }
 
-          self.connection = connection_class.new connection_options
-          self.client = Punchblock::Client.new :connection => connection
+          self.client = Punchblock.client_with_connection self.config.platform, connection_options
 
           # Tell the Punchblock connection that we are ready to process calls.
           Events.register_callback :after_initialized do
@@ -169,6 +162,10 @@ module Adhearsion
 
         def resource
           [Adhearsion::Process.fqdn, ::Process.pid].join '-'
+        end
+
+        def connection
+          client.connection
         end
       end
     end # Punchblock