+++ spec/adhearsion/call_controller/dial_spec.rb
@@ -245,7 +245,7 @@ module Adhearsion
             latch.wait(1).should be_false
 
             other_mock_call << mock_answered
-            other_mock_call << Punchblock::Event::Unjoined.new(:call_id => call.id)
+            other_mock_call << Punchblock::Event::Unjoined.new(call_uri: call.id)
 
             latch.wait(1).should be_false
 
@@ -389,7 +389,7 @@ module Adhearsion
 
             latch.wait
             time = Time.now - time
-            time.to_i.should be == timeout
+            time.round.should be == timeout
             t.join
             status = t.value
             status.result.should be == :timeout