+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/adhearsion/adhearsion)
+  * Feature: Add joined call attribute to dial status
   * Feature: Add Call#wait_for_end, which blocks until the call ends and returns its end reason
   * Bugfix: Removed unnecessary Mocha reference from generated plugin
   * Bugfix: Call loggers should be deleted after a call finishes