+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [develop](https://github.com/adhearsion/adhearsion)
   * Feature: Add Call#wait_for_end, which blocks until the call ends and returns its end reason
+  * Feature: Add joined call attribute to dial status
   * Bugfix: Removed unnecessary Mocha reference from generated plugin
   * Bugfix: Call loggers should be deleted after a call finishes