+++ CHANGELOG.md
@@ -2,7 +2,7 @@
   * Feature: Add Call#wait_for_end, which blocks until the call ends and returns its end reason
   * Feature: Add joined call attribute to dial status
   * Feature: Track call start/end time and duration
-  * Feature: Add per-call join duration to DialStatus
+  * Feature: Add per-call join duration and disposition to DialStatus
   * Bugfix: Removed unnecessary Mocha reference from generated plugin
   * Bugfix: Call loggers should be deleted after a call finishes