+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Feature: Add per-call join duration and disposition to DialStatus
   * Feature: Add `CallController#dial_and_confirm` which allows parallel confirmation of outbound calls
   * Feature: Add `#originate` method to console as alias for `Adhearsion::OutboundCall.originate`
+  * Feature: Allow the console to be disabled using `--no-console`
   * Bugfix: Removed unnecessary Mocha reference from generated plugin
   * Bugfix: Call loggers should be deleted after a call finishes